Use importlib_resources to import UI files
This is an example of using importlib_resources to import the
UI files. It now can use the UI file without having to run hexrd
in the same directory as the UI file.

It currently gives the QUiLoader a file path. If the python package
is zipped, it creates a temporary file to give the QUiLoader. It might
be better to figure out how to give QUiLoader the data directly,
perhaps in the form of QBuffer.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>