Store the UI-generated QMainWindow in MainWindow
It is easier to make connections between GUI and non-GUI objects
if set up like this. However, I am not sure if MainWindow still
actually needs to inherit QMainWindow now...

If a better way can be found in the future, where the UI file
objects get saved directly into the MainWindow object, then
perhaps we should go with that instead.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>