Added materials panel and "Add Material" dialog
This is a basic materials panel that allows you to choose a material,
and it displays the hkl values , d-spacing, and 2 theta. Selecting
a material will set the material to be used for calibration.

In addition, an "Add Material" dialog is present as well. It lets
the user choose the space group, lattice parameters, etc. But it
does not actually add a material yet.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>