Turn Configuration class into a Singleton
The Configuration class has been used as if it were a singleton for a
while, but it hasn't actually been a singleton. Officially turning
it into a singleton makes its usage easier.

It is also planned to store the loaded images in the configuration
class. This will save time so that the images do not have to be
reloaded for every operation.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>