Load images in config, and store them as dict
This loads the images once, and they are not re-loaded when
calibrating or toggling tab view. This hopefully will save
a little bit of time.

The keys for the dictionary are the instrument panel ids.
We do not yet have a way for the user to ensure that they
match, but we will add it soon. Currently, the names of the
files much match the instrument panel ids.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>