Added basic functionality to materials panel
This adds basic functionality to most of the components of the
materials panel. Rings can now be shown or hidden. The ranges
can be shown or hidden. The ranges can be specified with a spinbox.
And specific rings can be selected for viewing by clicking on their
hkl rows in the table.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>