Kill the program when ctrl-c is used
Because Qt catches and processes signals when it is
running, python does not immediately see the "ctrl-c" signal
when it is entered until Qt hands control back to the python
interpreter. This results in "ctrl-c" not working.

For some reason, adding the signal connection that is done in
this commit fixes the issue. The program now closes with "ctrl-c".
No saving is performed - just a hard terminate.

See [here](https://stackoverflow.com/questions/5160577/ctrl-c-doesnt-work-with-pyqt)
for related discussion.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>