Adding a new material now works properly
A material can now be added and used for the calibration.
It cannot yet be saved, though, and materials cannot be
modified either.

We may want to modify the lattice spinboxes because, currently,
when they are edited, they immediately do a lot of work which
slows down the program and can raise exceptions if the user is
in the process of modifying the spinboxes.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>