Update active material energy when needed
This causes the active material's energy to be updated when
it is needed. This is a potentially expensive operation that
can slow down many things, including start-up time, so it should
generally be avoided. But it is necessary to keep the material
in sync with the instrument energy.

This also fixes an issue with the default materials where, for
some reason, the beam energy of the material does not match the
beam energy of the material's plane data. This can cause issues
when calculating a new set of plane data, because the energy will
be different. This issue is also fixed in this commit by updating
the default materials' beam energy to match that of the plane data.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>