Added ability to modify materials
Users can modify materials by renaming them and changing their
parameters (including their space group, max hkl, and lattice
parameters).

Also, when a user adds a material, it now pops up with a blocking
dialog. This was done to prevent a clash with the dialog for
modifying a material, where the accepted signal would be
intercepted.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>