Include all hkl values for a material's PlaneData
When creating PlaneData for a material, the default behavior
is to exclude all hkl's after the fifth one. However, we allow
the user to choose the maxHKL, which is supposed to kind of serve
this purpose.

Therefore, get rid of all of the exclusions so that all hkls up
to the maxHKL are included.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>