Emit signals from HexrdConfig for big changes
As of this commit, the HexrdConfig() class now inherits from QObject,
and it emits signals when important changes occur.

The first application for this is also in this commit: when new plane
data is generated for the active material, a signal is emitted. This
signal is connected to the materials panel, and the materials panel
updates its contents immediately after the new plane data is generated.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>