Always fix exclusions when generating new pdata
Added a utils file for some generic operations. Now,
when new PlaneData is generated in the hexrd configuration,
the exclusions are fixed as well. This allows more than 5
rows to show up in the materials panel (and it also allows more
than 5 rings to appear in the detector view).

This also reverted a couple of changes - particularly, we no longer
copy the current material for the add material dialog, but we
just use the default material with a space group of 255.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>