Add live updates for materials panel changes
Now, when the user modifies the materials panel or the beam
energy, the rings in the canvas are automatically updated.
This is done by the configuration emitting a signal, the canvases
checking to see if they are currently displaying a calibration,
and then updating the rings if they are.

This also restructures the canvases slightly so that, when they
are displaying a calibration, they store in memory an
InstrumentViewer object, so they do not need to re-create a lot
of objects (such as the hexrd instrument object) when updating
the rings. It re-uses the InstrumentViewer to update the rings.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>