Use model/view approach for the tree view
This gives us much more explicit control over the tree
view. In particular, it allows us to prevent the "key"
column from being edited - which was an issue before.

It seems to work in most of my testing, but more testing
will certainly need to be done in the future.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>