Fix max hkl values in materials.hexrd
The max hkl values are by default really high, and
it causes tons of rings to appear if the beam energy
is modified, or if the material itself is modified.

This reduces the max hkl values and then saves over the
materials.hexrd file. The file size appears to be much
smaller as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>