Do not allow cached gui yaml paths to be edited
These gui yaml paths are cached when they are used so that
subsequent lookups are faster. We should return a deep copy
so that the cached paths will not be edited from the outside.

The cached path was edited when setting a value for a detector,
and this caused a bug that is now fixed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>