Turn off keyboard tracking for lattice params
If keyboard tracking is enabled (which it is by default for
spin boxes), signals get fired for every keystroke while the
user in the middle of editing. This can be very annoying if
it is doing validation and expensive operations each time, which
it is indeed doing for the lattice parameters.

Turn off keyboard tracking so that the expensive operations and
validation are performed AFTER the user finishes typing, not
during.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>