Added ability to display saturation percentages
When in image viewer mode, a user can now choose to display
saturation percentages in the bottom left corners of the images.

If a user checks the box, "View"->"Show Saturation Percentages",
saturation percentages will be displayed in white text in the
bottom left corner of every image. The saturation percentages are
calculated using the detector's "saturation_level" setting.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>