Allow saving of imageseries as hdf5 or npz
This allows users to save an imageseries as either an hdf5 file
or an npz file. The option is `Save`->`ImageSeries`.

The dialogs that appear go like this:

1. If there is more than one imageseries available, it will ask the
user to pick an imageseries via a QInputDialog combo box.

2. When the user is choosing the file to save to, the type (either
hdf5 or npz) is chosen based upon the filter the user selects. It is
NOT based upon any file endings the user gives the file.

3. If an NPZ file was chosen, the user then needs to pick a threshold
from a QInputDialog double spinbox. This is a required option for NPZ
in hexrd.

I am able to successfully modify an imageseries, save it as an HDF5
file, load it, and see that the modifications still remain. So HDF5
seems to work.

For NPZ, I can write an npz file, but then hexrd cannot load it for
some reason. I might need to change something to make it work.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>