Update instrument config names
A few instrument config names were changed in the latest updates
of hexrd (currently on the master branch [here](https://github.com/joelvbernier/hexrd3)).

The changes:
- `t_vec_d` -> `translation`
- `t_vec_s` -> `translation`
- `tilt_angles` -> `tilt`

Note that after this commit, an error will occur when starting hexrd
if previous QSettings are loaded. This can be prevented by the fix
in #81.

Also, running a default calibration is a tad bit slow because there are
many overlays. This is fixed by #83.

Otherwise, the calibrations now seem to work as they should.

Fixes: #82

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>