Run powder calibration
This adds a button that runs a powder calibration. It then updates
the instrument configuration with the new, optimized instrument
configuration. It then re-draws the canvas.

There are, however, several hard-coded values, including that all
flags are currently refinable. And some options need to be forwarded
to the user. These will be fixed soon.

It seems that the values need to be relatively close to their
optimized values in order for the optimization to work.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>