Allow user to choose Euler angle convention
This is a menu option under "Edit"->"Euler Angle Convention".
It brings up a combo box where the user chooses the convention
that they wish to use. This convention is then used for the
powder calibration (this is currently the only place where it is
used).

It is also saved in between runs.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>