For the cartesian view, rescale floating images
The adaptive histogram equalization requires floating point images
to be between -1.0 and 1.0. This rescales floating point images so
that is true.

It rescales it to be between 0.0 and 1.0 rather than -1.0 and 1.0
because, when we rescale it to be between -1.0 and 1.0, the images
will not show up in the GUI for some reason.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>