Always use imageseries for storing images
This allows us to be much more flexible for different kinds of
images. For instance, we can now do image processing on tiff files.

It loads the image files by creating a temporary yaml file that the
"image-files" imageseries class reads in.

It has a whitelist for the kinds of images that it will attempt to
load with the "image-files" imageseries. The current whitelist consists
of "tiff" only - more will need to be added if we wish to load more
using the "image-files" imageseries.

Deep copy would not work for "image-files" imageseries in
frame_aggregation.py and process_ims_dialog.py. However, I noticed that
the scripts do not edit the images, but only change the contents of
the dictionary. So a shallow copy works fine.

There is also now an index in the configuration that keeps track of
the active image series index.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>