Update tilt angles on Euler convention change
This updates the values of the tilt angles when the Euler
convention changes.

It seems to be working correctly for the most part. Switching
between Extrinsic and Intrinsic seems to work. Switch between
Extrinsic and None also seems to work. But switching between
Intrinsic and None results in some strange behavior.

Also, sometimes the tiles get oriented incorrectly when switching
from Intrinsic to Extrinsic, but they get re-oriented correctly
when switching back to Intrinsic. Additionally, when switching the
Euler angle convention, sometimes two of the tiles disappear in the
polar view.

In addition, the "Form View" now displays degrees when using any
Euler convention except for "None". The degrees automatically get
converted to radians before the configuration is edited.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>