Plot azimuthal integration in polar view
This is a start for plotting the azimuthal integration in the
polar view. I think it is working okay, although I am not sure
why the baseline isn't zero for my example.

There is white margin space on the left and right sides of the
images, which probably needs to be fixed. Also, the overlays
do not stretch all the way to the top and bottom of the
azimuthal integration plot, and hiding/showing the rings
expands the y axis for some reason.

Let me know how the actual plot looks.

Fixes: #122

@joelvbernier
@cryos

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>