Reload previous active material if available
When starting the program, reload the previous active material if
it is available.

This will currently just reload default materials (with all of their
default settings) that were previously selected. So, if ruby was
previously selected, for instance, it will automatically change the
active material to ruby when the program starts. But ruby will have
all of its default settings (none of the settings such as the max hkl
are currently saved in between runs).

Let me know if we want to save other settings in the config as well.

Addresses part of #107

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>