Reload all canvases when a new config is loaded
This adds an option to MainWindow.update_all() to allow for
clearing of all canvases. This ensures that they are all re-drawn
from scratch.

The commit also uses this option to re-draw all canvases when a
new config is loaded.

I noticed that if a new config is loaded and the normal
MainWindow.update_all() is called without clearing the canvases,
the rings in the Cartesian view are placed a little differently
than if the canvas is cleared. In my case, the rings were placed
a little more downward (in y) if the canvas was cleared beforehand.
I am not sure what is causing the difference, but I assume the
cleared canvas is probably the correct one, which is why we are
clearing the canvases.

If clearing the canvases is something we want to avoid when we
reload the config, I can try to look into what is causing the
difference.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>