Indicate x,y positions and intensity on status bar
When the mouse is moved over an image/plot, show the x and y
positions as well as the intensity on the status bar.

The intensity was unfortunately a little bit tricky to obtain,
because there did not seem to be a straightforward way in
matplotlib to get the intensity. However, I was able to find
an example online of someone else calculating it, and I used
that.

It is currently displaying the values in the bottom left corner,
because that is the default spot for temporary messages on the status
bar. Let me know if I should move it elsewhere.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>