Add a few processing updates to status bar
When running some processing steps that could take a while,
update the label in the bottom right corner of the status
bar.

The label update function also takes a timeout argument that,
when specified, will clear the label after the timeout
(in milliseconds) is completed. This is primarily for "finished"
labels such as "Powder calibration finished!".

Updates were added for loading image, Cartesian, and Polar views,
and for the powder calibration.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>