Always use temporary messages for the status bar
Rather than having a separate label that gets modified for processing
updates, always use temporary messages. Now, all status updates appear
in the bottom left corner.

In addition, the status now also gets cleared if the mouse moves over
the white part of the canvas (any part of the canvas that is not a
plot/image).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>