Set defaults for required missing config keys
This loops through several config settings (currently the
instrument, resolution, and calibration settings), and it
adds any keys that are present in the default config settings
but missing in the loaded config settings.

This will hopefully make the program more resilient to things
such as loading config files that have missing fields. Rather
than a key error occurring, it will automatically add in the
missing keys with the default values.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>