Added image mode widget, removed view buttons
The "image mode widget" replaces the "resolution widget". When the tab
is changed in the image mode widget, the image mode is changed. This
makes the buttons obsolete, and they were removed.

In addition to these things, the checkboxes for tabbed view and
showing the saturation level were moved out of the "View" menu in the
menu bar, and they were moved to the "Raw" tab of the image mode
widget.

There are a lot of changes introduced by this PR, some of which may
break other PRs. We should test this branch and investigate a little
more before merging.

Fixes: #132

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>