Added slider view to edit detector transforms
The slider view is a convenient way to edit the translation and tilt
transforms of the detectors. The user can choose a detector from the
combo box, and then modify the translations and tilts by either
moving the sliders or modifying the QDoubleSpinBoxes.

It works with live update. Currently, it causes one re-rendering for
every group of changes that are within 100 milliseconds of each other.

It would be ideal if we could find ways to speed up re-rendering. One
way, for instance, would be to only render the detector that is
currently being transformed. We will need to look into the code
to see if this is possible, though.

Other improvements are probably needed. For instance, the translation
ranges are currently fixed to -5000 and 5000, and the tilt ranges are
fixed to -180 and 180. We should perhaps come up with a way to make the
ranges smaller and modifiable in order to control the slider
sensitivity. Right now, relatively small movements of the sliders cause
relatively big changes in the rendering.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>