Re-render only the modified detector for cartesian
This speeds up cartesian view re-rendering significantly. The changes
are mostly intended for the slider widget currently, but will probably
be extended to the other widgets as well.

This is probably not bug free, and needs to be tested.

We will also need to speed up re-rendering for the polar view.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>