Fix reset range when old max is less than new min
Before this fix, if the default range was something like
"2181 - 5164", and the range was changed to "200 - 400", when
"Reset Range" would be clicked, the range values would appear
as "400 - 5164", and then, when clicked again, they would appear
as "2181 - 5164".

This is because the min value was being limited by the old max
value, 400 in this case. Ensure the new min is not limited by
the old max value.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>