Display omega tooltips, calculate max or delta
Require that minimum omega be input and calculate delta given maximum, or maximum
given delta. Hovering over a file will display a tooltip with the full file path,
the minimum column will explain that the min is requied, and the maximum and
delta columns will explain that one or the other must be input.

Signed-off-by: Brianna Major <brianna.major@kitware.com>