Emit valueChanged when scientific spinbox steps
Before, valueChanged() was not being emitted when the up/down arrows
were clicked. This resulted in the internal config not being updated.

Use setValue() on the spin box so that the signals get emitted. Also,
call "selectAll()" so that it behaves more like a regular spinbox,
which also calls "selectAll()" when the arrows are clicked.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>