Draw a border around detectors in polar view
A border that has the value of the maximum integer in the image
is drawn around the detector in the polar view. A large percentage (3%)
of the detector shape is required for it to show up in polar mode.

The borders in the cartesian view were also modified so that they take
up a percentage of the detector shape rather than an absolute value.
They were also modified to use the max integer in the image instead of
a percentile, since some detectors have very few pixels that are
illuminated.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>