Update logic for YAML files
In the case of raw images, prevent changing the number of empty frames (this
needs to be set or changed in the YAML file). Additionally, while the number of
empty frames should be displayed in the table it should not be used in
processing the image series when working with raw images, as the offset is taken
care of in the initial image series creation.

Signed-off-by: Brianna Major <brianna.major@kitware.com>