Don't grow azimuthal y-axis on ring selections
Because the azimuthal overlays are being drawn from the top of the
azimuthal plot to the bottom, the axis would get automatically
re-scaled so that there is a margin between the bounds of the image
and the overlays.

This means that every time the overlays would be re-drawn, the y-axis
would grow, and the plot would look smaller. This would happen every
time the user changed ring selections.

We can either fix this issue by turning off autoscaling for the
azimuthal plot, or by manually setting the y limits of the plot after
it has been autoscaled. I went with the former.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>