Update all when Euler convention is changed
In an ideal world, I don't think the canvases would need to be updated
when the Euler angle convention changed, because it should just be
another representation of the same information.

However, that is unfortunately not the case. I have especially noticed
that when going from None or Extrinsic (XYZ) to Intrinsic (ZXZ), there
is a difference in the images. In addition, if there is an angle such
as +180 degrees, there is a random chance it will be -180 degrees
instead, which changes the appearance of the detectors in the canvas.

Prior to this change, the user would only see the update when they
changed a value in the instrument config. But we want to make it clear
that the update is happening because of the change in Euler angle
convention, so update it immediately.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>