Update polar view on change of beam angle sliders
This updates the polar view every 500 milliseconds via a QTimer as the
beam angle sliders change.

Changing the beam angle modifies the entire image, so no speedup via
updating only one detector is possible. Because of this, the update is
not fast, and needs to be done every 500 milliseconds at most.

Changing the beam angle does not appear to have any effect on the
Cartesian view, so it is currently ignored.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>