Added calibration line picker
If the user chooses "Run"->"Calibration Line Picker" (which is
only enabled in polar mode), they enter a calibration line picking
mode, where they can left-click to add points to a line, and
right-click to start a new line. Once the user is done picking
points, they click "Ok" on the pop-up dialog box that is waiting,
and it sends the data they picked to the main window for a
calibration.

It currently does not do any calibrating, but just printing the data.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>