Move "Modify Material Dialog" into materials panel
This makes the "Modify Material Dialog" into a widget, and it moves
it into the materials panel.

When the values in the spinboxes are modified (keyboard tracking is
off), or the combo boxes change, the rings and the materials panel
table are updated immediately.

You can add a material or remove a material via the tool button.

You can also rename the material since the combo box is now editable.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>