Use "None" tilt convention to make HEDMInstrument
Before, the tilt angles in our instrument config were being passed
into the HEDMInstrument as the values that were modified by changing
the tilt convention.

Now, they are converted back to the "None" tilt convention before
being passed into the HEDMInstrument class.

Additionally, they are converted to the "None" tilt convention before
the detectors are updated for fast detector updates.

I believe this fixes all of the issues we were seeing with the Euler
angle convention.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>