Allow APS ImageSeries directory munging
To keep everything pertaining to file loading in the load panel, this adds the
ability to select entire directories to import.

To use this feature, each directory should be named for the detector it
represents. Select one of these directories with the 'Load Image File' dialog and
check the 'Use APS ImageSeries' checkbox (these may be done in any order), then
select the 'Load Image Files' button. When the 'Use APS ImageSeries' checkbox is
ticked, the 'Select Image Files' will not display a file selection dialog, but
will instead select all files contained within that directory.

This follows the logic used for 'Use Subdirectories' and expects that each
detector folder shares the same parent folder. The sibling directories are found,
and all files are loaded into the file table.

Signed-off-by: Brianna Major <brianna.major@kitware.com>