Add ability to show multiple materials
This adds the ability to plot all of the material overlays at the
same time. A checkbox was added to the materials panel that is
called "Show All Materials". If checked, all material overlays will
be plotted.

Additionally, there is now also an "Edit Style..." button on the
materials panel. With this, the ring and range styles of the current
material can be modified. The line colors can be set, along with the
line styles and the line width. These style settings are saved to the
config in between runs.

Being able to edit the line styles of the current material allow for
differentiation between materials when they are being simultaneously
plotted.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>