Add ability to modify material visibilities
A new checkbox was added to the materials panel that allows the user
to toggle on and off individual material visibilities. Now, only
materials that are flagged as "visible" will be displayed in the
main render window. However, multiple materials can be visible at
the same time. Each material's overlay styles can be modified by
selecting the material in the material panel, and then clicking
"Edit Style...".

If a row is selected in the hkl table, only the individual overlay
for the active material will be shown (and only if the active material
is currently visible - if not, nothing will be shown).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>