Allow transform and dark mode to be set per detector
This adds a checkbox that allows the user to choose whether to apply the selected
transformation and dark mode to all detectors, or to set them per detector. The
default state is per detector.

When selecting per detector, each detector must be changed by selecting the
detector from the drop down menu and then updating the tranformation and/or dark
mode. If left unchanged, that detector will receive the default value. Switching
between detectors will update values accordingly and make it clear which values
will be used for that detector when the files are read in.

When 'Apply Selections to All Detectors' has been selected, the current settings
will be used for all detectors, regardless of which detector is currently
visible. Switching between detectors will display the same transformation and
dark mode selections for all.

Signed-off-by: Brianna Major <brianna.major@kitware.com>