Add ring selection limits to materials panel
This allows a user to specify either a max Bragg angle
or a minimum d spacing to be used for ring selection. Any rings
that are higher than the max Bragg angle or lower than the min
d-spacing will not be displayed in the GUI.

It still does not yet update exclusions to be used for the powder
calibration. Please let me know if that is desired.

Addresses part of #107

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>