Use tThMax on plane data for max bragg angle
Rather than storing the max bragg angle inside hexrdgui's configuration,
now we set and get the max bragg angle via the tThMax on materials'
plane data. The max bragg angle can be calculated by dividing tThMax by
two, and the min d spacing can be calculated using Bragg's law.

This also allows each material to have its own max bragg angle, rather
than setting the same max bragg angle to all materials. Thus, the
materials panel now shows the max bragg angle and the min d spacing
for the current material, rather than for all materials.

Additionally, "Limit active", when toggled off, now sets the tThMax on
the material's plane data to None. When toggled on, it restores the
most recent tThMax for the material (which is not saved in between
runs, but only in the current run). This was done so that unchecking
and checking will not change the tThMax to something else.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>