Use plane data exclusions for ring selections
Rather than having a separate internal variable for the ring selections,
use the material's plane data exclusions instead.

This means that selecting rows now modifies the exclusions, and when
materials are loaded, rows are programmatically selected to match
the exclusions. The row selections are also now preserved when
changing materials. Thus, the row selections *should* always match
the exclusions of the currently selected material.

Modifying the exclusions also makes it so that, when a calibration is
performed, the rings that are used in the calibration are the selected
ones (the other rings get ignored).

Now, nearly all of the individual materials' properties on the materials
panel can be saved and loaded in the "materials.hexrd", including ring
selections, and the max bragg angle. The only properties that are not
being saved in the "materials.hexrd" file are:

1. Whether the material is visible or not
2. The overlay styles of the material
3. The range width (this is a global property for all materials
   currently)

Most of these are saved in the local hexrd config file in between runs.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>