Always clear saturation before another is shown
This fixes a problem where, if files are opened more than one time
during a single run, the old saturation text doesn't get cleared
before the new one gets placed on top of it.

This makes it so that the saturation text is always cleared before
the next saturation text gets drawn.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>