Plot detector borders on top of the image
Previously, if "View" -> "Show Detector Borders" was checked, the
cartesian and polar images would be modified, with the detector borders
drawn in as the max pixel values. This meant that the borders would
affect the azimuthal integration plot in the polar plot.

Rather than drawing the detector borders into the images, plot them as
separate lines on top of the images. This way, the images and the
azimuthal integration plot are not affected by their presence.

This seems to work well, although drawing the detector borders for the
polar plot is time consuming. We are using
[hexrd.instrument.PlanarDetector.pixel_angles()](https://github.com/HEXRD/hexrd/blob/79ebd31e43ca4b96a7cad60ff61ff2cc4fb14fba/hexrd/instrument.py#L1636)
for this, and it appears to be slowing down the generation of the polar
plot by a lot. It would be good to figure out a way to get the polar
detector borders faster. The `pixel_angles()` function generates and
returns a lot of values we don't need (we only extract the first and
last rows/columns from the output). Maybe we can make a new function
that only generates the detector border values.

The drawing of the Cartesian borders seems to be fast, and it remains
very interactive with the slider widgets.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>