Only convert border pixels to angles in polar view
For drawing the detector borders in the polar view, we were previously
converting all pixels to angles, and then extracting the border angles
from the results. This was slow, though, because so many pixels were
being converted to angles.

Now, we are obtaining only the border pixels, and then converting only
them into angles. The process of drawing the borders in the polar view
is now much faster, and the user is able to interact with the slider
widgets again (however, note that in both Cartesian and Polar views,
it will probably always be slightly faster to interact with the images
without borders being drawn).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>