Convert borders to degrees in a better location
For the dexela instrument, I would be getting this error when
trying to draw the polar borders:

```
AttributeError: 'numpy.ndarray' object has no attribute 'degrees'
```

I'm not sure what was causing the error, but it was happening at
`np.degrees(borders)` where borders was a list of tuples of
np.ndarray objects.

It seems that the error does not occur if we convert to degrees
earlier, so let's just do that instead.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>