Allow each material its own tThWidth
Rather than forcing all materials to have the same tThWidth,
as was done previously, allow each material to have their own
tThWidths. The "Show Ranges" checkbox was renamed to "Enable Width",
and this toggles setting `None` on the material's tThWidth. The
spinbox displays the tThWidth on the material. And, when the active
material in the materials panel is changed, the checkbox and spinbox
update to the newly selected material.

This also changes a little naming in the materials panel, including
renaming "Show Rings" to a more general "Show Overlays".

Additionally, the plus or minus prefix on the width was removed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>