Make sure config dicts get cleared upon clear
Copies of these config dicts are sometimes saved elsewhere
in the code. When we clear the config dicts with "self.var = {}",
the copies of the config dicts elsewhere in the code may not update
their state and keep a copy of the old config dict around, and work
with it.

If we call "self.var.clear()" on the config dict instead, it ensures
that the dictionary is cleared, and all shallow copies of the config
dict elsewhere in the code will be modified as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>