Fix image tab widget bug
This fixes the bug mentioned
[here](https://github.com/HEXRD/hexrdgui/issues/261), where it appears
that drawing the dummy images before the event loop has started causes
a weird bug where, when using "Tabbed View" in the image tab widget,
the first two tabs do not respond when the user clicks on them.

There is also a timer being used in an attempt to avoid a black render
flash.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>