Draw dummy images after first paint event
Drawing the dummy images after the first paint event fixes two issues:

1. A black window that occurs when the program first starts up, because
   the program has not finished drawing the dummy images when it tries
   to show().
2. A bug where the "Tabbed View" in the image tab widget doesn't work
   correctly if the images are drawn before the event loop has started
   (see https://github.com/HEXRD/hexrdgui/issues/261).

This solution is better than the previous one, because it doesn't
depend on a QTimer being timed to ~100 ms. It appears to work well
when it is posted to the event loop immediately after the first
paint event.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>