Update images when detectors are added/removed
This causes the dummy images to be updated when detectors
are removed/added. It also adds a new signal in the hexrd
config that causes a deep re-render to occur.

Currently, there are many things that can cause a deep re-render
to occur, but most of them occur due to internal checks in the
ImageCanvas class (such as "_polar_reset_needed()"). This adds
the ability for other parts of the code to force a deep re-render
as well. It is needed, for instance, if the user switches to a
config that has the same number of detectors (because the ImageCanvas
class will not automatically perform a deep re-render in this case).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>