Add ability to export polar plot to HDF5
When the polar plot is being displayed, the user can export the
polar plot data via "File"->"Export"->"Polar Plot". Now, the user may
choose between HDF5 (default) or npz for the export. If the file
name ends with ".npz", a ".npz" file will be written. Otherwise,
an HDF5 file will be written.

The data that is written out is as follows:
```
tth_coordinates
eta_coordinates
intensities
extent
```

They are all written out as datasets in the root in the HDF5 file.

Currently, no material overlays are written out and no borders are
written out. But that is potentially changeable.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>